+++ opensalestax-woocommerce.php
@@ -4,7 +4,7 @@
  * Plugin Name:       OpenSalesTax for WooCommerce
  * Plugin URI:        https://github.com/ejosterberg/opensalestax-woocommerce
  * Description:       Calculate US sales tax via a self-hosted OpenSalesTax engine. Replaces TaxJar / Avalara / WooCommerce Tax integrations with a free, open-source alternative.
- * Version:           0.4.1
+ * Version:           0.5.0
  * Requires at least: 6.2
  * Requires PHP:      8.2
  * Author:            Eric Osterberg